+++ src/main/kotlin/com/example/backend/controller/PlanController.kt
@@ -1,6 +1,8 @@
 package com.example.backend.controller
 
-import com.example.backend.domain.models.*
+import com.example.backend.domain.models.response.*
+import com.example.backend.domain.models.typeEnum.ImageType
+import com.example.backend.domain.models.typeEnum.PlanType
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
@@ -9,11 +11,11 @@ import java.util.*
 @RestController
 class PlanController {
     @GetMapping("api/v1/plan/top")
-    fun getTopPlan(): List<TopPlan> {
-        val res = mutableListOf<TopPlan>()
+    fun getTopPlan(): List<TopPlanResponse> {
+        val res = mutableListOf<TopPlanResponse>()
 
         // 1つ目
-        val sampleTopPlan1 = TopPlan()
+        val sampleTopPlan1 = TopPlanResponse()
         sampleTopPlan1.id = 1
         sampleTopPlan1.title = "Sample Plan 1"
         sampleTopPlan1.image = "none"
@@ -27,7 +29,7 @@ class PlanController {
         res += sampleTopPlan1
 
         // 2つ目
-        val sampleTopPlan2 = TopPlan()
+        val sampleTopPlan2 = TopPlanResponse()
         sampleTopPlan2.id = 2
         sampleTopPlan2.title = "Sample Plan 2"
         sampleTopPlan2.image = "none"
@@ -42,7 +44,7 @@ class PlanController {
         res += sampleTopPlan2
 
         // 3つ目
-        val sampleTopPlan3 = TopPlan()
+        val sampleTopPlan3 = TopPlanResponse()
         sampleTopPlan3.id = 3
         sampleTopPlan3.title = "Sample Plan 3"
         sampleTopPlan3.image = "none"
@@ -60,11 +62,11 @@ class PlanController {
     }
 
     @GetMapping("api/v1/plan/info")
-    fun getPlanInfo(@RequestParam id: Int) : List<PlanInfoElement> {
-        val res = mutableListOf<PlanInfoElement>()
+    fun getPlanInfo(@RequestParam id: Int) : List<PlanInfoResponse> {
+        val res = mutableListOf<PlanInfoResponse>()
 
         if(id in 1..3){
-            val planInfo1 = PlanInfoElement()
+            val planInfo1 = PlanInfoResponse()
             planInfo1.id = 1
             planInfo1.title = "移動"
             planInfo1.body = "車"
@@ -74,7 +76,7 @@ class PlanController {
             planInfo1.type = PlanType.MOVE.id
             res += planInfo1
 
-            val planInfo2 = PlanInfoElement()
+            val planInfo2 = PlanInfoResponse()
             planInfo2.id = 2
             planInfo2.title = "海"
             planInfo2.body = "海"
@@ -84,7 +86,7 @@ class PlanController {
             planInfo2.type = PlanType.SPOT.id
             res += planInfo2
 
-            val planInfo3 = PlanInfoElement()
+            val planInfo3 = PlanInfoResponse()
             planInfo3.id = 3
             planInfo3.title = "昼食"
             planInfo3.body = "海の家で昼食"
@@ -94,7 +96,7 @@ class PlanController {
             planInfo3.type = PlanType.SPOT.id
             res += planInfo3
 
-            val planInfo4 = PlanInfoElement()
+            val planInfo4 = PlanInfoResponse()
             planInfo4.id = 4
             planInfo4.title = "移動"
             planInfo4.body = "車"
@@ -104,7 +106,7 @@ class PlanController {
             planInfo4.type = PlanType.MOVE.id
             res += planInfo4
 
-            val planInfo5 = PlanInfoElement()
+            val planInfo5 = PlanInfoResponse()
             planInfo5.id = 4
             planInfo5.title = "観光"
             planInfo5.body = "美術館"
@@ -114,7 +116,7 @@ class PlanController {
             planInfo5.type = PlanType.SPOT.id
             res += planInfo5
 
-            val planInfo6 = PlanInfoElement()
+            val planInfo6 = PlanInfoResponse()
             planInfo6.id = 4
             planInfo6.title = "移動"
             planInfo6.body = "帰宅"
@@ -129,7 +131,7 @@ class PlanController {
     }
 
     @GetMapping("/api/v1/plan/detail")
-    fun getPlanInfoDetail(@RequestParam id: Int): PlanDetailResponse{
+    fun getPlanInfoDetail(@RequestParam id: Int): PlanDetailResponse {
         var res = PlanDetailResponse()
         res.id = 1
         res.title = "観光"
@@ -163,5 +165,6 @@ class PlanController {
 
         return res
     }
+
 
 }
