+++ src/main/kotlin/com/example/backend/domain/models/response/ImageClass.kt
@@ -1,4 +1,4 @@
-package com.example.backend.domain.models
+package com.example.backend.domain.models.response
 
 data class ImageClass (
     var image_type: Int? = null,
