+++ src/main/kotlin/com/example/backend/domain/models/typeEnum/ImageType.kt
@@ -1,4 +1,4 @@
-package com.example.backend.domain.models
+package com.example.backend.domain.models.typeEnum
 
 enum class ImageType(val id: Int) {
     MAIN(1),
