+++ src/main/kotlin/com/example/backend/domain/models/typeEnum/PlanType.kt
@@ -1,4 +1,4 @@
-package com.example.backend.domain.models
+package com.example.backend.domain.models.typeEnum
 
 enum class PlanType(val id: Int) {
     SPOT(1),
